requirements.txt
# Django
django
# Django Rest Framework
djangorestframework
markdown
django-filter